% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angdiff.R
\name{angdiff}
\alias{angdiff}
\title{Computes differences between two angles}
\usage{
angdiff(x, y)
}
\arguments{
\item{x}{A numeric value in [0;360] representing an angle.}

\item{y}{A numeric value in [0;360] representing an angle.}
}
\value{
a numeric value representing the difference angle between \code{x}
and \code{y}.
}
\description{
\code{angdiff} computes the difference between two angles.
}
\examples{
#
}
\seealso{
.
}
