% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herdersmovementpatterns-package.R
\docType{package}
\name{herdersmovementpatterns-package}
\alias{herdersmovementpatterns}
\alias{herdersmovementpatterns-package}
\title{herdersmovementpatterns: Research compendium for "Patterns in Mongolian Nomadic Household Movement Derived from GPS Trajectories"}
\description{
This is the research compendium \insertCite{Marwick.2019}{herdersmovementpatterns} for the
  manuscript "Patterns in Mongolian Nomadic Household Movement Derived from GPS Trajectories", including the
  Rmarkdown script, and data used to generate the manuscript. No raw trajectory data ar provided for ethical
  reasons.
}
\author{
\strong{Maintainer}: Henning Teickner \email{henning.teickner@uni-muenster.de}

Authors:
\itemize{
  \item Christian Knoth \email{christian.knoth@uni-muenster.de}
  \item Bartoschek Thomas
  \item Kati Kraehnert
  \item Melinda Vigh
  \item Myagmartseren Purevtseren
  \item Munkhnaran Sugar
  \item Kati Kraehnert
  \item Edzer Pebesma
}

}
\keyword{internal}
