% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_altitude_difference.R
\name{mp_altitude_difference}
\alias{mp_altitude_difference}
\title{Computes the altitude difference between the highest and lowest point of a Track object.}
\usage{
mp_altitude_difference(t)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{altitude} in the data slot.}
}
\value{
a numeric value representing the altitude difference between the
highest and the lowest point in \code{t}.
}
\description{
\code{mp_altitude_difference} computes the altitude difference between
the highest and the lowest point of a
\code{\link[trajectories:Track-class]{Track}} object with a variable
\code{altitude} in the data slot. \code{NA}s are removed.
}
\examples{
#
}
\seealso{
.
}
