% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_altitude_total_relative_distance.R
\name{mp_altitude_total_relative_distance}
\alias{mp_altitude_total_relative_distance}
\title{Computes the total altitude distance covered by a Track object.}
\usage{
mp_altitude_total_relative_distance(t)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{altitude} in the \code{data} slot.}
}
\value{
a numeric value representing the total altitude distance covered by
\code{t}, relative to the total horizontal distance covered by t.
}
\description{
\code{mp_altitude_total_relative_distance} computes the total altitude distance
covered by a \code{Track} (sum of absolute altitude
distances) divided by the total horizontal distance covered by the same track.
}
\examples{
#
}
\seealso{
.
}
