% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_chull.R
\name{mp_chull}
\alias{mp_chull}
\title{Extracts the convex hull from a Track object.}
\usage{
mp_chull(t)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}.}
}
\value{
A \code{\link[sp:SpatialPolygons]{SpatialPolygons}}
with coordinates corresponding to coordinates of points of the convex hull.
}
\description{
\code{mp_chull} extracts the convex hull from a
\code{Track} object and returns it
as \code{SpatialPolygons}.
}
\examples{
#
}
\seealso{
.
}
