% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_distance_total.R
\name{mp_distance_total}
\alias{mp_distance_total}
\title{Computes the total distance covered by a Track object.}
\usage{
mp_distance_total(t)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}.}
}
\value{
a numeric value representing the total distance covered by \code{t}.
}
\description{
\code{mp_distance_total} computes the total distance covered by a
\code{Track}. This is the sum of all distance
values in the \code{connections} slot.
}
\examples{
#
}
\seealso{
.
}
