% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_gap_proportion.R
\name{mp_gap_proportion}
\alias{mp_gap_proportion}
\title{Computes the proportion of gaps in a Track object.}
\usage{
mp_gap_proportion(t)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location}.}
}
\value{
A numeric value representing the proportion of gaps
in \code{t}.
}
\description{
\code{mp_gap_proportion} computes the proportion of gaps in a
\code{Track}.
}
\examples{
#
}
\seealso{
.
}
