% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_latitude_difference.R
\name{mp_latitude_difference}
\alias{mp_latitude_difference}
\title{Computes the latitude difference between the most northern and most southern points of a Track object.}
\usage{
mp_latitude_difference(t)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{altitude} in the data slot.}
}
\value{
a numeric value representing the latitudinal difference between the
most northern and most southern points in \code{t}.
}
\description{
\code{mp_latitude_difference} computes the latitudinal difference between
the most northern and most southern points of a
\code{Track}.
}
\examples{
#
}
\seealso{
.
}
