% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_latitude_sum.R
\name{mp_latitude_sum}
\alias{mp_latitude_sum}
\title{Computes the sum of the latitudinal transitions of a Track object.}
\usage{
mp_latitude_sum(t)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite} in the data slot.}
}
\value{
a numeric value representing the sum of the latitude transitions
for \code{t}.
}
\description{
\code{mp_latitude_sum} sums the latitude differences between adjacent campsite
locations for a \code{Track}.
}
\examples{
#
}
\seealso{
.
}
