% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_locations_count.R
\name{mp_locations_count}
\alias{mp_locations_count}
\title{Counts unique locations of a Track object.}
\usage{
mp_locations_count(t, campsite = TRUE, repeated = FALSE)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location}, \code{campsite} and \code{norepeatedcampsitevisits}.
as created by \code{\link[herdersTA]{locationsTrack}} in the data slot.}

\item{campsite}{A logical value indicating if only campsites should be considered
(\code{TRUE}) or any locations (\code{FALSE}).}

\item{repeated}{A logical value indicating if the number of locations with repeated
visits shoud be counted (\code{TRUE}) or not (\code{FALSE}). This is only available
with \code{campsite = TRUE}.}
}
\value{
a numeric value representing the number of unique (campsite) locations
in \code{t}.
}
\description{
\code{mp_locations_count} counts unique lcoations of a
\code{Track}. Several options are available.
}
\examples{
#
}
\seealso{
.
}
