% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_locations_direction_average.R
\name{mp_locations_direction_average}
\alias{mp_locations_direction_average}
\title{Computes the average direction (as absolute difference of angles) between campsite locations of a Track object in the order of movement.}
\usage{
mp_locations_direction_average(t, fun = stats::median)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite}}

\item{fun}{One of \code{mean} or \code{median)}, depending on which function should be
used in order to compute average values.}
}
\value{
a numeric value representing the average direction between
locations in \code{t} in the order of movement.
}
\description{
\code{mp_locations_direction_average} computes the average direction between
locations of a \code{\link[trajectories:Track-class]{Track}} in the order of movement.
Directions are computed as angle difference between the previous campsite location
and the current campsite location and the next campsite location and the current
campsite location. Several options are available.
}
\examples{
#
}
\seealso{
.
}
