% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_longitude_difference.R
\name{mp_longitude_difference}
\alias{mp_longitude_difference}
\title{Computes the longitude difference between the most western and most eastern points of a Track object.}
\usage{
mp_longitude_difference(t)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}.}
}
\value{
a numeric value representing the longitudinal difference between the
most western and most eastern points in \code{t}.
}
\description{
\code{mp_longitude_difference} computes the longitude difference between
the most western and most eastern points of a
\code{Track}.
}
\examples{
#
}
\seealso{
.
}
