% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_longitude_sum.R
\name{mp_longitude_sum}
\alias{mp_longitude_sum}
\title{Computes the sum of the longitude transitions of a Track object.}
\usage{
mp_longitude_sum(t)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite} in the data slot.}
}
\value{
a numeric value representing the sum of the longitude transitions
for \code{t}.
}
\description{
\code{mp_longitude_sum} sums the longitude differences between adjacent campsite
locations for a \code{\link[trajectories:Track-class]{Track}}.
}
\examples{
#
}
\seealso{
.
}
