% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_number_sequence.R
\name{mp_number_sequence}
\alias{mp_number_sequence}
\title{Nicely prints sequences of numbers.}
\usage{
mp_number_sequence(x, digits = 2)
}
\arguments{
\item{x}{A numeric vector with at least two elements.}

\item{digits}{An integer vale representing the digit to which the
elements of \code{x} should be rounded.}
}
\value{
A string representing \code{x} as nicely formatted sequence.
}
\description{
\code{mp_number_sequence} takes a numeric vector and nicely prints its elements
as sequence.
}
\examples{
x <- c(23.345, 232.4234, 12.3)
mp_number_sequence(x, digits = 1)
}
