% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_pca_biplot_scores.R
\name{mp_pca_biplot_scores}
\alias{mp_pca_biplot_scores}
\title{Creates biplots of the loadings of a PCA for different combinations of PCs.}
\usage{
mp_pca_biplot_scores(scores_df, choices = seq_len(ncol(scores_df)))
}
\arguments{
\item{scores_df}{A \code{matrix} of loadings of a PCA as returned by
\code{\link[vegan:scores]{scores}} with the score values.}

\item{choices}{A numeric vector of column indices in \code{loadings_df} with the
indices of PCs for which to create plots.}
}
\value{
An object of class \code{\link[ggplot2:ggplot]{ggplot2}} with
a panel for each combination of input PCs showing the scores for each sample.
}
\description{
\code{mp_pca_biplot_scores} takes a \code{matrix} of scores of a PCA
as returned by \code{\link[vegan:scores]{scores}} and creates a biplot of the scores for
each combination of a set of selected PCs.
}
