% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_pca_biplot_variables.R
\name{mp_pca_biplot_variables}
\alias{mp_pca_biplot_variables}
\title{Creates biplots of the loadings of a PCA for different combinations of PCs.}
\usage{
mp_pca_biplot_variables(
  loadings_df,
  choices = seq_len(ncol(loadings_df)),
  x_limits = c(-2, 1),
  c = 0.05
)
}
\arguments{
\item{loadings_df}{A \code{matrix} of loadings of a PCA as returned by
\code{\link[vegan:scores]{scores}} with the loading values.}

\item{choices}{A numeric vector of column indices in \code{loadings_df} with the
indices of PCs for which to create plots.}

\item{x_limits}{A numeric vector with two elements indicating the x axis limits
of the plots to create.}

\item{c}{A numeric value representing a multiplication factor that indicates by which
amount the length of an arrow is increased to place the variable label.}
}
\value{
An object of class \code{\link[ggplot2:ggplot]{ggplot2}} with
a panel for each combination of input PCs showing the loadings for each variable.
}
\description{
\code{mp_pca_biplot_variables} takes a \code{matrix} of loadings of a PCA
as returned by \code{\link[vegan:scores]{scores}} and creates a biplot of the loadings for
each combination of a set of selected PCs.
}
