% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_si.R
\name{mp_si}
\alias{mp_si}
\title{Computes various summary indicators for a Track object.}
\usage{
mp_si(t, fun = stats::median)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite}}

\item{fun}{One of \code{mean} or \code{median)}, depending on which function should be
used in order to compute average values.}
}
\value{
a numeric value representing the number of unique (campsite) locations
in \code{t}.
}
\description{
\code{mp_si} is a wrapper function for functions that takes a \code{Track}
object as input and return numeric values/vectors as output that summarise the
\code{Track} object. The functions are:
\enumerate{
  \item \code{\link{mp_locations_altitude_average}}
  \item \code{\link{mp_altitude_range}} (to compute the minimum and maximum altitude)
  \item \code{\link{mp_altitude_difference}}
  \item \code{\link{mp_altitude_total_distance}}
  \item \code{\link{mp_altitude_total_relative_distance}}
  \item \code{\link{mp_locations_distance_average}}
  \item \code{\link{mp_distance_total}}
  \item \code{\link{mp_longitude_difference}}
  \item \code{\link{mp_latitude_difference}}
  \item \code{\link{mp_longitude_sum}}
  \item \code{\link{mp_latitude_sum}}
  \item \code{\link{mp_chull}} (to compute the area and perimeter of the convex hull)
  \item \code{\link{mp_locations_direction_average}}
  \item \code{\link{mp_locations_direction_sd}}
  \item \code{\link{mp_locations_direction_average}}
  \item \code{\link{mp_locations_count}} (to count unique campsite locations and repeated campsite visits)
  \item \code{\link{mp_gap_proportion}}
  \item \code{\link{mp_straightness_index}}
  \item \code{\link{mp_linearity}}
}
}
\examples{
#
}
\seealso{
.
}
