% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_si_pca_variable_contribution.R
\name{mp_si_pca_variable_contribution}
\alias{mp_si_pca_variable_contribution}
\title{Computes the variable contribution for a PCA}
\usage{
mp_si_pca_variable_contribution(loadings)
}
\arguments{
\item{loadings}{A \code{matrix} of loadings of a PCA as returned by
\code{\link[vegan:scores]{scores}} with the loading values.}
}
\value{
A \code{data.frame} with one row for each variable used in the PCA
and three columns:
\describe{
  \item{variable_name}{A character vector with the variable names.}
  \item{PC}{The principal component.}
  \item{contribution}{A numeric vector with the contribution of the variables.}
}
}
\description{
\code{mp_si_pca_variable_contribution} takes a \code{matrix} of loadings of a PCA
as returned by \code{\link[vegan:scores]{scores}} and computes the contribution of the
individual variables used in the PCA.
}
