% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_si_summarise.R
\name{mp_si_summarise}
\alias{mp_si_summarise}
\title{Computes the mean, median, minimum, maximum, and number of samples in a vector.}
\usage{
mp_si_summarise(x, digits = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{An integer vale representing the digit to which the
elements of \code{x} should be rounded.}
}
\value{
A \code{data.frame} with one row and a column for each the
mean (mean), median (median), minimum (min), maximum (max), and number of samples (n).
}
\description{
\code{mp_si_summarise} takes a numeric vector and computes its mean, median, minimum,
maximum, and number of non-\code{NA} elements, and rounds the results.
}
\examples{
x <- c(23.345, 232.4234, 12.3, 23423.23, 456, 23)
mp_si_summarise(x, digits = 1)
}
