% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_straightness_index.R
\name{mp_straightness_index}
\alias{mp_straightness_index}
\title{Computes the straightness index for the trajectory between the first and last campsite location in a Track object.}
\usage{
mp_straightness_index(t)
}
\arguments{
\item{t}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite}}
}
\value{
a numeric value representing the straightness index for
\code{t}.
}
\description{
\code{mp_straightness_index} computes the straightness index
\insertCite{Laube.2007}{herdersmovementpatterns} for the
trajectory between the first and last locations of a
\code{Track} object in the order of movement.
}
\examples{
#
}
\references{
\insertAllCited{}
}
\seealso{
.
}
