% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tc_rbind.R
\name{mp_tc_rbind}
\alias{mp_tc_rbind}
\title{Binds the data slots of \code{Track} objects in a \code{Trackscollection} object row-wise.}
\usage{
mp_tc_rbind(tc)
}
\arguments{
\item{tc}{A \code{\link[trajectories:Track-class]{TracksCollection}} object
with all \code{Track} objects having the same length and temporal resolution.}
}
\value{
a \code{data.table} object with thw row-wise binded data slots of the
\code{Track} objects of \code{tc} and two additional variables, \code{household}
(a numeric vector with an id for each household) and \code{time} (a \code{Date}
vector with the date information of each \code{Track} object).
}
\description{
\code{mp_tc_rbind} binds the data slots of \code{\link[trajectories:Track-class]{Track}} objects
in a \code{\link[trajectories:Track-class]{TracksCollection}} object row-wise
}
\examples{
#
}
\seealso{
.
}
