% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tcdf_extract_gaps.R
\name{mp_tcdf_extract_gaps}
\alias{mp_tcdf_extract_gaps}
\title{Extracts information on gaps from a \code{data.frame} representing a \code{Trackscollection} object.}
\usage{
mp_tcdf_extract_gaps(tcdf, dummy = TRUE)
}
\arguments{
\item{tcdf}{A \code{data.frame} as returned by \code{\link{mp_tc_rbind}}
if applied on a \code{\link[trajectories:Track-class]{TracksCollection}}
object.}

\item{dummy}{A logical value indicating if in the case that for a household
no gap is detected, a dummy entry should be inserted in the resulting
\code{data.frame} (\code{dummy = TRUE}) or not (\code{dummy = FALSE}).}
}
\value{
a \code{data.frame} object with the following variables:
\describe{
  \item{\code{household}}{A numeric vector with an id for each household.}
  \item{\code{location}}{A numeric vector with an id for each location per
  household.}
  \item{\code{group}}{A numeric vector with an id for each visit per
  household and location.}
  \item{\code{row_start}}{A numeric vector with an id for the row of
  \code{tcdf}, where a gap has its first value.}
  \item{\code{row_end}}{A numeric vector with an id for the row of
  \code{tcdf}, where a gap has its last value.}
  \item{\code{time_start}}{A \code{Date} vector with the start date of
  each gap.}
  \item{\code{time_end}}{A \code{Date} vector with the end date of
  each gap.}
  \item{\code{duration}}{A numeric vector with the duration of each gap
  in days.}
  \item{\code{is_first}}{A logical vector indicating for each gap if it is
  the first gap for a household.}
  \item{\code{is_last}}{A logical vector indicating for each gap if it is
  the last gap for a household.}
  \item{\code{is_gap}}{A logical vector indicating if the current row represents
  a gap (always \code{TRUE}).}
  \item{\code{location_previous}}{A numeric vector with an id for the location
  prior the gap.}
  \item{\code{location_next}}{A numeric vector with an id for the location
  after the gap.}
  \item{\code{moved}}{A logical vector indicating for each gap if a movement
  occured during the gap. This is always the case if
  \code{location_previous != location_next}.}
  \item{\code{proportion}}{A numeric vector with the temporal proportion of each gap
  relative to the duration of the trajectory of the household.}
}
}
\description{
\code{mp_tcdf_extract_gaps} extracts information on gaps from a
\code{\link[trajectories:Track-class]{TracksCollection}} object
that was combined using \code{\link{mp_tc_rbind}}.
}
\examples{
#
}
\seealso{
.
}
