% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_clip.R
\name{tc_clip}
\alias{tc_clip}
\title{Clips \code{TrackCollections} to specified time windows}
\usage{
tc_clip(tc, d)
}
\arguments{
\item{tc}{An object of class
\code{\link[trajectories:Track-class]{TracksCollection}} with one
\code{Track} per \code{Tracks} and the same length and time resolution
for each \code{Track} object.}

\item{d}{A data.frame with a row for each time window and two columns:
\describe{
  \item{\code{start}}{A \code{Date} vector with the start time for each
  time window.}
  \item{\code{end}}{A \code{Date} vector with the end time for each
  time window.}
}}
}
\value{
A list with one \code{{TracksCollection}} for each row in \code{d}
representinga clipped form of \code{tc}.
}
\description{
\code{tc_clip} clips the \code{\link[trajectories:Track-class]{Track}} objects of
a \code{\link[trajectories:Track-class]{TracksCollection}} object for
specified time windows.
}
\examples{
#
}
\seealso{
.
}
